% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichrdb.R
\name{getEnrichrdb}
\alias{getEnrichrdb}
\title{Get EnrichrDB geneset and geneset_name
Geneset is a data.frame of 2 columns with term id and gene id}
\usage{
getEnrichrdb(
  org = c("human", "fly", "yeast", "worm", "zebrafish"),
  library = NULL,
  download.method = NULL
)
}
\arguments{
\item{org}{Organism from 'human','fly','yeast','worm','zebrafish'.}

\item{library}{Choose one library name from `enrichr_metadata`.}

\item{download.method}{"auto" (as default if NULL), "wininet" (for windows)}
}
\value{
A list including geneset and geneset name.
}
\description{
Get EnrichrDB geneset and geneset_name
Geneset is a data.frame of 2 columns with term id and gene id
}
\examples{
\dontrun{
x = getEnrichrdb(org = "human", library = "COVID-19_Related_Gene_Sets")
}
}
