% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGO.R
\name{getGO}
\alias{getGO}
\title{Get GO geneset and geneset_name
Geneset is a data.frame of 2 columns with term id and gene id;
Geneset_name is a data.frame of 2 columns with term id and term description}
\usage{
getGO(org = "human", ont = c("bp", "cc", "mf"), download.method = NULL)
}
\arguments{
\item{org}{Organism name from `go_org`.}

\item{ont}{One of "bp","cc" or "mf" ontology.}

\item{download.method}{"auto" (as default if NULL), "wininet" (for windows)}
}
\value{
A list including geneset and geneset name.
}
\description{
Get GO geneset and geneset_name
Geneset is a data.frame of 2 columns with term id and gene id;
Geneset_name is a data.frame of 2 columns with term id and term description
}
\examples{
\dontrun{
x = getGO(org = "human",ont = "mf")
}
}
