% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHgDisease.R
\name{getHgDisease}
\alias{getHgDisease}
\title{Get HgDisease geneset and geneset_name
Human disease gene sets from Disease Ontology (DO),DisGeNET,
Network of Cancer Gene (NCG) version 6 and v7, and covid19-specific.
Geneset is a data.frame of 2 columns with term id and gene id;
Geneset_name is a data.frame of 2 columns with term id and term description}
\usage{
getHgDisease(
  source = c("do", "disgenet", "ncg_v7", "ncg_v6", "covid19"),
  download.method = NULL,
  data_dir = NULL
)
}
\arguments{
\item{source}{Choose from 'do','ncg_v7','ncg_v6','disgenet','covid19'.}

\item{download.method}{"auto" (as default if NULL), "wininet" (for windows)}

\item{data_dir}{data saving location, default is the package data directory}
}
\value{
A list.
}
\description{
Get HgDisease geneset and geneset_name
Human disease gene sets from Disease Ontology (DO),DisGeNET,
Network of Cancer Gene (NCG) version 6 and v7, and covid19-specific.
Geneset is a data.frame of 2 columns with term id and gene id;
Geneset_name is a data.frame of 2 columns with term id and term description
}
\examples{
\donttest{
x = getHgDisease(source = "do", data_dir = tempdir())
}
}
