% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKEGG.R
\name{getKEGG}
\alias{getKEGG}
\title{Get KEGG geneset and geneset_name
Geneset is a data.frame of 2 columns with term id and gene id;
Geneset_name is a data.frame of 2 columns with term id and term description}
\usage{
getKEGG(
  org = "hsa",
  category = c("pathway", "module", "enzyme", "disease", "drug", "network"),
  download.method = NULL
)
}
\arguments{
\item{org}{Organism name from `kegg_org`.}

\item{category}{Choose one category from "pathway","module", "enzyme",
'disease' (human only),'drug' (human only) or 'network' (human only)}

\item{download.method}{"auto" (as default if NULL), "wininet" (for windows)}
}
\value{
A list including geneset and geneset name.
}
\description{
Get KEGG geneset and geneset_name
Geneset is a data.frame of 2 columns with term id and gene id;
Geneset_name is a data.frame of 2 columns with term id and term description
}
\examples{
\dontrun{
x = getKEGG(org = "hsa",category = "pathway")
}
}
