% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReactome.R
\name{getReactome}
\alias{getReactome}
\title{Get Reactome geneset and geneset_name
Geneset is a data.frame of 2 columns with term id and gene id;
Geneset_name is a data.frame of 2 columns with term id and term description}
\usage{
getReactome(org = "human", download.method = NULL, data_dir = NULL)
}
\arguments{
\item{org}{Organism name from `reactome_org`.}

\item{download.method}{"auto" (as default if NULL), "wininet" (for windows)}

\item{data_dir}{data saving location, default is the package data directory}
}
\value{
A list including geneset and geneset name.
}
\description{
Get Reactome geneset and geneset_name
Geneset is a data.frame of 2 columns with term id and gene id;
Geneset_name is a data.frame of 2 columns with term id and term description
}
\examples{
\donttest{
x = getReactome(org = "human", data_dir = tempdir())
}
}
