% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWiki.R
\name{getWiki}
\alias{getWiki}
\title{Get WikiPathway geneset and geneset_name
Geneset is a data.frame of 2 columns with term id and gene id;
Geneset_name is a data.frame of 2 columns with term id and term description}
\usage{
getWiki(org = "human", download.method = NULL)
}
\arguments{
\item{org}{Organism name from `wiki_org`.}

\item{download.method}{"auto" (as default if NULL), "wininet" (for windows)}
}
\value{
A list including geneset and geneset name.
}
\description{
Get WikiPathway geneset and geneset_name
Geneset is a data.frame of 2 columns with term id and gene id;
Geneset_name is a data.frame of 2 columns with term id and term description
}
\examples{
\dontrun{
x = getWiki(org = "human")
}
}
