% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_gene_single.R
\name{GSEA_gene_single}
\alias{GSEA_gene_single}
\title{Do GSEA for a single geneSet based on random permuted or bootstraps by genes}
\usage{
GSEA_gene_single(gene_values = rnorm(100), gene_names = 1:100,
  this_geneSet = sample(1:100, 10), geneSet_name = "geneset1", n = 1000)
}
\arguments{
\item{gene_values}{A vector of values for each gene,most of time it's  signal 2 noise.}

\item{gene_names}{A vector of names  for each gene}

\item{this_geneSet}{A vector which contains all of the genes in this geneSet.}

\item{n}{Times for random permuted or bootstraps by genes,default:1000}

\item{this_geneSet}{The name of current geneSet,default:geneset1}
}
\value{
a list contain 4 elements: ES,rand_es, p and ismem
}
\description{
It'll do GSEA Tests for a single geneSet according to the signal 2 noise values for all of the genes or other values,such as foldchange.
It'll return ES ,the running ES score for all of the genes, a numeric vector .
rand_es ,the maximal running ES score for each permutation, a numeric vector.
p which calculated by formule:p <- 1-pnorm((max(ES)-mean(rand_es))/sd(rand_es))
ismem is a logical vector shows whether the gene belong to the current geneSet or not.
}
\examples{
GSEA_gene_single()
}
\keyword{GSEA}

