% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset.R
\name{gset}
\alias{gset}
\title{Build a \code{gset} object}
\usage{
gset(gsnames, gsdesc, gslist)
}
\arguments{
\item{gsnames}{character vector of gene set names}

\item{gsdesc}{character vector of gene set descriptions}

\item{gslist}{list of character vectors containing gene names}
}
\value{
a \code{gset} object. See above
}
\description{
Builds an object of class \code{gset}. Basically a list with the following fields:
\describe{
 \item{names}{character vector of gene set names}
 \item{descriptions}{character vector of gene set descriptions}
 \item{genesets}{list of character vectors containing gene symbols for each gene set}
}
}

