% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsintersect.R
\name{gsintersect}
\alias{gsintersect}
\title{Gene set/gene list intersect}
\usage{
gsintersect(gslist, genenames)
}
\arguments{
\item{gslist}{A list of gene sets (character vectors of gene symbols)}

\item{genenames}{A vector of gene symbols that you want to keep in \code{gslist}.}
}
\value{
A list of gene sets
}
\description{
Returns the intersection between each gene set in a gene set list and
a list of gene names. Useful for stripping gene sets of genes that aren't
in your own data.
}

