% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperGTest_multiple.R
\name{hyperGTest_multiple}
\alias{hyperGTest_multiple}
\title{Hypergeometric Tests for a list of geneSets}
\usage{
hyperGTest_multiple(geneSet_list = list(set1 = sample(1:100, 10), set2 =
  sample(1:100, 10)), choose_geneIds = sample(1:100, 20),
  backgroud_geneIds = 1:100)
}
\arguments{
\item{geneSet_list}{A list which contains all of the vector for each geneSet.}

\item{choose_geneIds}{A vector which contains all of gene which defined by user, probably significantly differently expression genes.}

\item{backgroud_geneIds}{A vector which contains all of the genes have been detected by microarray or RNA-seq, and so on}
}
\value{
a data.frame which has 8 columns:geneSet_name,N,n,M,k,exp_count,OddsRatio,p)
}
\description{
It'll do Hypergeometric Tests for a list of geneSets, such as KEGG database. And we need 3 parameters, first is this_geneSet which contains all of the genes in this geneSet.
Then is choose_geneIds,which probably is the significantly differently expression genes. Last is the backgroud_geneIds, which should be all of the genes which have been detected by microarray or RNA-seq, and so on .
The N is the size of backgroud_geneIds
The n is the size of choose_geneIds
The M is the size of intersect(this_geneSet,backgroud_geneIds)
The K is the size of intersect(this_geneSet,choose_geneIds)
The exp_count is n*M/N
The OddsRatio is k/exp_count
The p is the statistic P value for this Hypergeometric Tests
}
\examples{
hyperGTest_multiple()
}
\keyword{hyperGtest}

