% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperGTest_single.R
\name{hyperGTest_single}
\alias{hyperGTest_single}
\title{Hypergeometric Tests for a single geneSet}
\usage{
hyperGTest_single(this_geneSet = sample(1:100, 10),
  choose_geneIds = sample(1:100, 20), backgroud_geneIds = 1:100)
}
\arguments{
\item{this_geneSet}{A vector which contains all of the genes in this geneSet.}

\item{choose_geneIds}{A vector which contains all of gene which defined by user, probably significantly differently expression genes.}

\item{backgroud_geneIds}{A vector which contains all of the genes have been detected by microarray or RNA-seq, and so on}
}
\value{
a vector include(N,n,M,k,exp_count,OddsRatio,p)
}
\description{
It'll do Hypergeometric Tests for a single geneSet, such as DNA replication. And we need 3 parameters, first is this_geneSet which contains all of the genes in this geneSet.
Then is choose_geneIds,which probably is the significantly differently expression genes. Last is the backgroud_geneIds, which should be all of the genes which have been detected by microarray or RNA-seq, and so on .
The N is the size of backgroud_geneIds
The n is the size of choose_geneIds
The M is the size of intersect(this_geneSet,backgroud_geneIds)
The K is the size of intersect(this_geneSet,choose_geneIds)
The exp_count is n*M/N
The OddsRatio is k/exp_count
The p is the statistic P value for this Hypergeometric Tests
}
\examples{
hyperGTest_single()
}
\keyword{hyperGtest}

