% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immunological}
\alias{immunological}
\title{Blood immunological gene sets from humans}
\format{A \code{\link{gset}} object with 508 elements.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp#C7}
}
\usage{
immunological
}
\description{
Contains sets from the immunological collection (C7) of MSigDB. Sets contain the
word "blood" in their description and have homo sapiens in their organism
field.
}
\seealso{
\code{\link{gset}}
}
\keyword{datasets}
