% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gmt.R
\name{load_gmt}
\alias{load_gmt}
\title{Read a \code{.gmt} file}
\usage{
load_gmt(file)
}
\arguments{
\item{file}{the name of the file you want to read}
}
\value{
An object of class \code{gset} with the following elements:
 \item{names}{a character vector containing gene set names}
 \item{descriptions}{a character vector containing gene set descriptions}
 \item{genesets}{a list of character vectors containing gene symbols}
}
\description{
Reads a file of the Gene Matrix Transposed file format.
}

