% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_gene_multiple.R
\name{PCA_gene_multiple}
\alias{PCA_gene_multiple}
\title{Do PCA for a single geneSet based on  bootstraps by genes}
\usage{
PCA_gene_multiple(prefix = "test", exprSet, geneSet_list, n = 1000)
}
\arguments{
\item{prefix}{The prefix for all of the output files.( we don't need it now actually,just placeholder)}

\item{exprSet}{Matrix for microarray expression values,rownames must be genes, colnames must be samples}

\item{geneSet_list}{A list which contains all of the vector for each geneSet.}

\item{n}{Times for random permuted or bootstraps by genes,default:1000}
}
\value{
A numeric vector of P value for each PCA test about all of the geneSet .
}
\description{
It'll random choose the same number of gene with the size of this geneSet to extract the expression matrix and then do PCA for this small expression matrix .
}
\examples{
PCA_gene_multiple(exprSet=exprSet, geneSet_list=geneSet_list)
}
\keyword{PCA}

