% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_gene_single.R
\name{PCA_gene_single}
\alias{PCA_gene_single}
\title{Do PCA for a single geneSet based on  bootstraps by genes}
\usage{
PCA_gene_single(prefix = "test", exprSet, group_list, this_geneSet,
  n = 1000)
}
\arguments{
\item{prefix}{The prefix for all of the output files.( we don't need it now actually,just placeholder)}

\item{exprSet}{Matrix for microarray expression values,rownames must be genes, colnames must be samples}

\item{group_list}{Factors for two groups that are tested for differential expression.}

\item{this_geneSet}{A vector which contains all of the genes in this geneSet.}

\item{n}{Times for random permuted or bootstraps by genes,default:1000}
}
\value{
The p vaule for this PCA test
}
\description{
It'll random choose the same number of gene with the size of this geneSet to extract the expression matrix and then do PCA for this small expression matrix .
}
\examples{
PCA_gene_single(exprSet=exprSet,group_list=group_list, this_geneSet=this_geneSet)
}
\keyword{PCA}

