% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoking}
\alias{smoking}
\title{Gene expression signatures of cigarette smoking}
\format{A \code{\link{gset}} object with the following gene sets:
\describe{
 \item{Top 25 genes current vs. never smokers}{}
 \item{Top 25 genes former vs. never smokers}{}
}}
\source{
Huan T, Joehanes R, Schurmann C, Schramm K, Pilling LC, Peters MJ, et al.
A Whole-Blood Transcriptome Meta-Analysis Identifies Gene Expression Signatures of
Cigarette Smoking. Hum Mol Genet. 2016 Aug 29
}
\usage{
smoking
}
\description{
A dataset containing gene sets from Huan et al., 2016, see source.
}
\examples{
data(smoking)
smoking[1]
}
\seealso{
\code{\link{gset}}
}
\keyword{datasets}
