% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gmt.R
\name{write_gmt}
\alias{write_gmt}
\title{create GMT file}
\usage{
write_gmt(geneSet = kegg2symbol_list, gmt_file = "kegg2symbol.gmt")
}
\arguments{
\item{gmt_file}{The filename for gmt file.}

\item{kegg2symbol_list}{A list contains all of the geneSet, each geneSet has some genes.(HUGO symbol)}
}
\value{
write gmt file
}
\description{
create gmt format file for GSEA software by broad institute.
}
\examples{
#' write_gmt()
}
\keyword{gmt}

