% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Datasets}
\alias{Datasets}
\alias{go_org}
\alias{kegg_org}
\alias{wiki_org}
\alias{msigdb_org}
\alias{enrichr_metadata}
\title{Datasets
go_org contains GO species information}
\description{
Datasets
go_org contains GO species information

Datasets
kegg_org contains KEGG species information

Datasets
wiki_org contains WikiPathway species information

Datasets
msigdb_org contains Msigdb species information

Datasets
enrichr_metadata contains Enrichrdb information
}
\keyword{datasets}
