% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsfilter.R
\name{gsfilter}
\alias{gsfilter}
\title{Filter gene set list by size of set}
\usage{
gsfilter(geneset, min = 5, max = Inf)
}
\arguments{
\item{geneset}{A list of gene sets (character vectors)}

\item{min}{the smallest size of gene set to keep in the list}

\item{max}{the largest size of gene set to keep in the list (defaults to
infinity)}
}
\value{
A list of gene sets
}
\description{
Removes gene sets smaller than \code{min}, larger than \code{max} from the input
gene set list.
}

