% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsintersect.R
\name{gsintersect}
\alias{gsintersect}
\title{Gene set/gene list intersect}
\usage{
gsintersect(geneset, genenames)
}
\arguments{
\item{geneset}{a \code{gset} object}

\item{genenames}{A vector of gene symbols that you want to keep in \code{geneset}.}
}
\value{
a \code{gset} object
}
\description{
Returns the set intersection between each gene set in a \code{gset} and
a list of gene names. Useful for stripping gene sets of genes that aren't
in your own data.
}
\seealso{
\code{\link{gset}}
}

