% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSEA_single.R
\name{plot_GSEA_single}
\alias{plot_GSEA_single}
\title{draw 3 figures for the results of GSEA_gene_single function}
\usage{
plot_GSEA_single(GSEA_single_results, gene_values = rnorm(100),
  prefix = "test", imageType = "png")
}
\arguments{
\item{GSEA_single_results}{A list contain 4 elements: ES,rand_es, p and ismem}

\item{gene_values}{A vector of values for each gene,most of time it's  signal 2 noise.}

\item{prefix}{The prefix for the out images,default:test}

\item{imageType}{choose png,pdf,emf,tiff,jpeg,bmp}
}
\value{
3 figures:
}
\description{
Firstly you need to do GSEA Tests for a list of geneSets,and get the results, then just use this function to draw some pictures.
}
\examples{
plot_GSEA_single();lapply(GSEA_gene_multiple(),plot_GSEA_single)
}
\keyword{GSEA,plot}

